Round manager, tienda, victoria, minimo comportamiento de enemigos
De nuevo, para testear con la nave y poder moverla se tiene que arrancar desde el menu y darle a jugar
Si queres agregar algun nuevo tipo de enemigo, me vas a tener que consultar porque me la complique un toque je

+++ Assets/Enemys/EnemyGlobal/Enemy_movement.cs
@@ -8,11 +8,11 @@ public class Enemy_movement
     Vector3 _direction;
     EnemyGlobalScript _enemies;
     Rigidbody _rb;
-    int _speed;
+    float _speed;
     Transform _transform;
     float _speedRotation = 55f;
 
-    public Enemy_movement(Ship ship , EnemyGlobalScript enemies , Rigidbody rb , int speed , Transform transform)
+    public Enemy_movement(Ship ship , EnemyGlobalScript enemies , Rigidbody rb , float speed , Transform transform)
     {
         _ship = ship;
         _enemies = enemies;
